+++ internal/version/version.go
@@ -4,7 +4,10 @@ Copyright (c) Edgeless Systems GmbH
 SPDX-License-Identifier: AGPL-3.0-only
 */
 
-package versions
+/*
+Package version provides functionality to parse and process semantic versions, as they are used in multiple components of Constellation.
+*/
+package version
 
 import (
 	"encoding/json"
